+++ src/Middleware/DocBlockAnnotations.php
@@ -6,14 +6,17 @@ namespace JsonMapper\Middleware;
 
 use JsonMapper\Builders\PropertyBuilder;
 use JsonMapper\Enums\Visibility;
-use JsonMapper\Helpers\AnnotationHelper;
 use JsonMapper\JsonMapperInterface;
+use JsonMapper\ValueObjects\AnnotationMap;
 use JsonMapper\ValueObjects\PropertyMap;
+use JsonMapper\ValueObjects\PropertyType;
 use JsonMapper\Wrapper\ObjectWrapper;
 use Psr\SimpleCache\CacheInterface;
 
 class DocBlockAnnotations extends AbstractMiddleware
 {
+    private const DOC_BLOCK_REGEX = '/@(?P<name>[A-Za-z_-]+)(?:[ \t]+(?P<value>.*?))?[ \t]*\r?$/m';
+
     /** @var CacheInterface */
     private $cache;
 
@@ -22,8 +25,12 @@ class DocBlockAnnotations extends AbstractMiddleware
         $this->cache = $cache;
     }
 
-    public function handle(\stdClass $json, ObjectWrapper $object, PropertyMap $propertyMap, JsonMapperInterface $mapper): void
-    {
+    public function handle(
+        \stdClass $json,
+        ObjectWrapper $object,
+        PropertyMap $propertyMap,
+        JsonMapperInterface $mapper
+    ): void {
         $propertyMap->merge($this->fetchPropertyMapForObject($object));
     }
 
@@ -38,25 +45,18 @@ class DocBlockAnnotations extends AbstractMiddleware
 
         foreach ($properties as $property) {
             $name = $property->getName();
-            $docblock = $property->getDocComment();
+            $propertyDetails = $this->derivePropertyDetailsFromReflectionProperty($property);
 
-            if ($docblock === false) {
+            if (is_null($propertyDetails)) {
                 continue;
-            }
-
-            $annotations = AnnotationHelper::parseAnnotations($docblock);
-            $type = $annotations['var'][0];
-            $isArray = substr($type, -2) === '[]';
-            if ($isArray) {
-                $type = substr($type, 0, -2);
             }
 
             $property = PropertyBuilder::new()
                 ->setName($name)
-                ->setType($type)
-                ->setIsNullable(AnnotationHelper::isNullable($annotations['var'][0]))
+                ->setType($propertyDetails->getType())
+                ->setIsNullable($propertyDetails->isNullable())
                 ->setVisibility(Visibility::fromReflectionProperty($property))
-                ->setIsArray($isArray)
+                ->setIsArray($propertyDetails->isArray())
                 ->build();
             $intermediatePropertyMap->addProperty($property);
         }
@@ -64,5 +64,53 @@ class DocBlockAnnotations extends AbstractMiddleware
         $this->cache->set($object->getName(), $intermediatePropertyMap);
 
         return $intermediatePropertyMap;
+    }
+
+    private function derivePropertyDetailsFromReflectionProperty(\ReflectionProperty $property): ?PropertyType
+    {
+        $docBlock = $property->getDocComment();
+        if ($docBlock === false) {
+            return null;
+        }
+
+        $annotations = self::parseDocBlockToAnnotationMap($docBlock);
+
+        if (! $annotations->hasVar()) {
+            return null;
+        }
+
+        $type = $annotations->getVar();
+
+        $isArray = substr($type, -2) === '[]';
+        if ($isArray) {
+            $type = substr($type, 0, -2);
+        }
+
+        $nullable = stripos('|' . $type . '|', '|null|') !== false;
+
+        return new PropertyType($type, $nullable, $isArray);
+    }
+
+    public static function parseDocBlockToAnnotationMap(string $docBlock): AnnotationMap
+    {
+        // Strip away the start "/**' and ending "*/"
+        if (strpos($docBlock, '/**') === 0) {
+            $docBlock = substr($docBlock, 3);
+        }
+        if (substr($docBlock, -2) === '*/') {
+            $docBlock = substr($docBlock, 0, -2);
+        }
+        $docBlock = trim($docBlock);
+
+        $var = null;
+        if (preg_match_all(self::DOC_BLOCK_REGEX, $docBlock, $matches)) {
+            for ($x = 0, $max = count($matches[0]); $x < $max; $x++) {
+                if ($matches['name'][$x] === 'var') {
+                    $var = $matches['value'][$x];
+                }
+            }
+        }
+
+        return new AnnotationMap($var ?: null, [], null);
     }
 }