+++ src/Middleware/MiddlewareLogicInterface.php
@@ -7,8 +7,14 @@ namespace JsonMapper\Middleware;
 use JsonMapper\JsonMapperInterface;
 use JsonMapper\ValueObjects\PropertyMap;
 use JsonMapper\Wrapper\ObjectWrapper;
+use stdClass;
 
 interface MiddlewareLogicInterface
 {
-    public function handle(\stdClass $json, ObjectWrapper $object, PropertyMap $propertyMap, JsonMapperInterface $mapper): void;
+    public function handle(
+        stdClass $json,
+        ObjectWrapper $object,
+        PropertyMap $propertyMap,
+        JsonMapperInterface $mapper
+    ): void;
 }