+++ tests/Unit/JsonMapperTest.php
@@ -83,7 +83,7 @@ class JsonMapperTest extends TestCase
     /**
      * @covers \JsonMapper\JsonMapper
      */
-    public function testPushedMiddleareIsInvokedWhenMappingObject(): void
+    public function testPushedMiddlewareIsInvokedWhenMappingObject(): void
     {
         $jsonMapper = new JsonMapper(new PropertyMapper());
         $jsonMapper->push($this->middleware);
@@ -104,5 +104,61 @@ class JsonMapperTest extends TestCase
         $jsonMapper->mapObject(new \stdClass(), new \stdClass());
 
         self::assertTrue($this->middleware->isCalled());
+    }
+
+    /**
+     * @covers \JsonMapper\JsonMapper
+     */
+    public function testRemovedMiddlewareIsNotInvokedWhenMappingObject(): void
+    {
+        $jsonMapper = new JsonMapper(new PropertyMapper());
+        $jsonMapper->push($this->middleware);
+        $jsonMapper->remove($this->middleware);
+
+        $jsonMapper->mapObject(new \stdClass(), new \stdClass());
+
+        self::assertFalse($this->middleware->isCalled());
+    }
+
+    /**
+     * @covers \JsonMapper\JsonMapper
+     */
+    public function testRemovedByNameMiddlewareIsNotInvokedWhenMappingObject(): void
+    {
+        $jsonMapper = new JsonMapper(new PropertyMapper());
+        $jsonMapper->push($this->middleware, __METHOD__);
+        $jsonMapper->removeByName(__METHOD__);
+
+        $jsonMapper->mapObject(new \stdClass(), new \stdClass());
+
+        self::assertFalse($this->middleware->isCalled());
+    }
+
+    /**
+     * @covers \JsonMapper\JsonMapper
+     */
+    public function testShiftedMiddlewareIsNotInvokedWhenMappingObject(): void
+    {
+        $jsonMapper = new JsonMapper(new PropertyMapper());
+        $jsonMapper->unshift($this->middleware, __METHOD__);
+        $jsonMapper->shift();
+
+        $jsonMapper->mapObject(new \stdClass(), new \stdClass());
+
+        self::assertFalse($this->middleware->isCalled());
+    }
+
+    /**
+     * @covers \JsonMapper\JsonMapper
+     */
+    public function testPoppedMiddlewareIsNotInvokedWhenMappingObject(): void
+    {
+        $jsonMapper = new JsonMapper(new PropertyMapper());
+        $jsonMapper->push($this->middleware, __METHOD__);
+        $jsonMapper->pop();
+
+        $jsonMapper->mapObject(new \stdClass(), new \stdClass());
+
+        self::assertFalse($this->middleware->isCalled());
     }
 }