+++ tests/Unit/ValueObjects/PropertyMapTest.php
@@ -7,6 +7,7 @@ namespace JsonMapper\Tests\Unit\ValueObjects;
 use JsonMapper\Enums\Visibility;
 use JsonMapper\ValueObjects\Property;
 use JsonMapper\ValueObjects\PropertyMap;
+use JsonMapper\ValueObjects\PropertyType;
 use PHPUnit\Framework\TestCase;
 
 class PropertyMapTest extends TestCase
@@ -16,7 +17,7 @@ class PropertyMapTest extends TestCase
      */
     public function testPropertyCanBeAdded(): void
     {
-        $property = new Property('name', 'string', true, Visibility::PUBLIC());
+        $property = new Property('name', new PropertyType('string', true, false), Visibility::PUBLIC());
         $map = new PropertyMap();
         $map->addProperty($property);
 
@@ -40,7 +41,7 @@ class PropertyMapTest extends TestCase
      */
     public function testMapReturnsCorrectIterator(): void
     {
-        $property = new Property('name', 'string', true, Visibility::PUBLIC());
+        $property = new Property('name', new PropertyType('string', true, false), Visibility::PUBLIC());
         $map = new PropertyMap();
         $map->addProperty($property);
         $iterator = $map->getIterator();
@@ -55,13 +56,13 @@ class PropertyMapTest extends TestCase
     public function testCanBeConvertedToJson(): void
     {
         $map = new PropertyMap();
-        $map->addProperty(new Property('id', 'int', false, Visibility::PUBLIC()));
+        $map->addProperty(new Property('id', new PropertyType('int', false, false), Visibility::PUBLIC()));
 
         $mapAsJson = json_encode($map);
 
         self::assertIsString($mapAsJson);
         self::assertJsonStringEqualsJsonString(
-            '{"properties":{"id":{"name":"id","type":"int","isNullable":false,"visibility":"public","isArray":false}}}',
+            '{"properties":{"id":{"name":"id","type":{"type":"int","isNullable":false,"isArray":false},"visibility":"public"}}}',
             (string) $mapAsJson
         );
     }
@@ -72,13 +73,13 @@ class PropertyMapTest extends TestCase
     public function testCanBeConvertedToString(): void
     {
         $map = new PropertyMap();
-        $map->addProperty(new Property('id', 'int', false, Visibility::PUBLIC()));
+        $map->addProperty(new Property('id', new PropertyType('int', false, false), Visibility::PUBLIC()));
 
         $mapAsString = $map->toString();
 
         self::assertIsString($mapAsString);
         self::assertJsonStringEqualsJsonString(
-            '{"properties":{"id":{"name":"id","type":"int","isNullable":false,"visibility":"public","isArray":false}}}',
+            '{"properties":{"id":{"name":"id","type":{"type":"int","isNullable":false,"isArray":false},"visibility":"public"}}}',
             (string) $mapAsString
         );
     }
@@ -89,9 +90,9 @@ class PropertyMapTest extends TestCase
     public function testCanBeMergedWithOtherPropertyMap(): void
     {
         $map = new PropertyMap();
-        $map->addProperty(new Property('id', 'int', false, Visibility::PUBLIC()));
+        $map->addProperty(new Property('id', new PropertyType('int', false, false), Visibility::PUBLIC()));
         $other = new PropertyMap();
-        $other->addProperty(new Property('uuid', 'string', false, Visibility::PUBLIC()));
+        $other->addProperty(new Property('uuid', new PropertyType('string', false, false), Visibility::PUBLIC()));
 
         $map->merge($other);